切换成实时计算模式

+++ animationer/src/main/java/com/base/animation/cache/PathCache.kt
@@ -12,15 +12,17 @@ class PathCache {
     /**
      * 根据每帧时间和路径组成唯一的key去判断路径是否一样
      */
-    fun conventKey(intervalDeal: Long, animPath: AnimPathObject): String {
+    fun conventKey(intervalDeal: Float, animPath: AnimPathObject): String {
         val startPoints = animPath.startPoints.toMutableMap()
         val endPointsMap = animPath.animPathMap.toMutableMap()
         var key = "$intervalDeal"
         startPoints.map {
             key += "_"
             key += it.value.map {
-                "start_${it.alpha}_${it.point.x}_${it.point.y}_${it.rotation}_${it.scaleX}_${it
-                    .scaleY}_${it.interpolator.javaClass.name}"
+                "start_${it.alpha}_${it.point.x}_${it.point.y}_${it.rotation}_${it.scaleX}_${
+                    it
+                        .scaleY
+                }_${it.interpolator.javaClass.name}"
             }
         }
         endPointsMap.map {
@@ -29,8 +31,10 @@ class PathCache {
                 val pathObject = it.pathObject
                 "end_${pathObject.alpha}_${pathObject.point.x}_${pathObject.point.y}_" +
                         "${pathObject.rotation}_${pathObject.scaleX}_${pathObject.scaleY}" +
-                        "_${pathObject.interpolator
-                    .javaClass.name}"
+                        "_${
+                            pathObject.interpolator
+                                .javaClass.name
+                        }"
             }
         }
         return key
